I added: The paging in GetFolderContentFromId() method, file upload/download, module refactoring.
