+++ Decisions.MSOneDrive/Decisions.MSOneDrive.csproj
@@ -73,12 +73,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Class1.cs" />
+    <Compile Include="Data\OneDrivePermission.cs" />
     <Compile Include="Data\OneDriveResource.cs" />
     <Compile Include="Data\OneDriveResult.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utility\AuthenticationHelper.cs" />
-    <Compile Include="Utility\CoreUtility.cs" />
-    <Compile Include="Utility\ModuleUtility.cs" />
+    <Compile Include="Utility\OneDriveFolders.cs" />
+    <Compile Include="Utility\OneDriveUtility.cs" />
+    <Compile Include="Utility\OneDriveFiles.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
