+++ Decisons.MSOneDrive.TestSuite/Decisons.MSOneDrive.TestSuite.csproj
@@ -94,10 +94,9 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ModuleTest.cs" />
+    <Compile Include="ModuleTests.cs" />
     <Compile Include="TestAuthenticationHelper.cs" />
     <Compile Include="TestData.cs" />
-    <Compile Include="CoreTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
